Create the react-hooks package
