+++ packages/online-editor/src/companionFs/CompanionFsHooks.ts
@@ -15,7 +15,8 @@
  */
 
 import { CompanionFsService, CompanionFsServiceBroadcastEvents } from "./CompanionFsService";
-import { Holder, useCancelableEffect } from "../reactExt/Hooks";
+import { Holder } from "@kie-tools-core/react-hooks/dist/Holder";
+import { useCancelableEffect } from "@kie-tools-core/react-hooks/dist/useCancelableEffect";
 import { useCallback } from "react";
 import {
   WORKSPACES_BROADCAST_CHANNEL,
@@ -23,7 +24,7 @@ import {
   WorkspacesBroadcastEvents,
   WorkspacesFilesBroadcastEvents,
 } from "@kie-tools-core/workspaces-git-fs/dist/worker/api/WorkspacesBroadcastEvents";
-import { usePromiseState } from "@kie-tools-core/workspaces-git-fs/dist/hooks/PromiseState";
+import { usePromiseState } from "@kie-tools-core/react-hooks/dist/PromiseState";
 import { decoder } from "@kie-tools-core/workspaces-git-fs/dist/encoderdecoder/EncoderDecoder";
 
 export function useSyncedCompanionFs(companionFsService: CompanionFsService) {
