+++ packages/online-editor/src/dates/RelativeDate.tsx
@@ -18,7 +18,7 @@ import moment from "moment";
 import * as React from "react";
 import { useCallback, useState } from "react";
 import { Tooltip } from "@patternfly/react-core/dist/js/components/Tooltip";
-import { useCancelableEffect } from "../reactExt/Hooks";
+import { useCancelableEffect } from "@kie-tools-core/react-hooks/dist/useCancelableEffect";
 
 export function RelativeDate(props: { date: Date }) {
   const [dateToDisplay, setDateToDisplay] = useState(moment(props.date).fromNow());
