+++ packages/online-editor/src/editor/DmnRunner/DmnRunnerDrawerPanelContent.tsx
@@ -33,7 +33,9 @@ import {
   InputRow,
   extractDifferences,
 } from "@kie-tools/form-dmn";
-import { Holder, useCancelableEffect, usePrevious } from "../../reactExt/Hooks";
+import { Holder } from "@kie-tools-core/react-hooks/dist/Holder";
+import { useCancelableEffect } from "@kie-tools-core/react-hooks/dist/useCancelableEffect";
+import { usePrevious } from "@kie-tools-core/react-hooks/dist/usePrevious";
 import { ErrorBoundary } from "../../reactExt/ErrorBoundary";
 import { EmptyState, EmptyStateBody, EmptyStateIcon } from "@patternfly/react-core/dist/js/components/EmptyState";
 import { I18nWrapped } from "@kie-tools-core/i18n/dist/react-components";
