+++ packages/online-editor/src/previewSvgs/PreviewSvgHooks.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Holder } from "../reactExt/Hooks";
+import { Holder } from "@kie-tools-core/react-hooks/dist/Holder";
 import { useCallback } from "react";
 import { decoder } from "@kie-tools-core/workspaces-git-fs/dist/encoderdecoder/EncoderDecoder";
 import { useCompanionFsFileSyncedWithWorkspaceFile } from "../companionFs/CompanionFsHooks";
