+++ packages/react-hooks/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@kie-tools-core/workspaces-git-fs",
+  "name": "@kie-tools-core/react-hooks",
   "version": "0.0.0",
   "description": "",
   "license": "Apache-2.0",
@@ -23,18 +23,8 @@
     "test": "run-script-if --ignore-errors \"$(build-env tests.ignoreFailures)\" --bool \"$(build-env tests.run)\" --then \"jest --silent --verbose --passWithNoTests\""
   },
   "dependencies": {
-    "@kie-tools-core/envelope-bus": "workspace:*",
-    "@kie-tools-core/react-hooks": "workspace:*",
-    "@kie-tools-core/workspace": "workspace:*",
-    "@kie-tools/emscripten-fs": "^0.0.2",
-    "@kie-tools/kie-sandbox-fs": "workspace:*",
-    "isomorphic-git": "^1.11.1",
-    "jszip": "^3.7.1",
-    "minimatch": "^3.0.4",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^5.2.1",
-    "uuid": "^8.3.2"
+    "react-dom": "^17.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.16.0",
@@ -47,13 +37,10 @@
     "@testing-library/react": "^11.2.6",
     "@types/jest": "^26.0.23",
     "@types/jest-when": "^2.7.4",
-    "@types/minimatch": "^3.0.4",
     "@types/react": "^17.0.6",
     "@types/react-dom": "^17.0.5",
-    "@types/react-router": "^5.1.14",
     "@types/testing-library__jest-dom": "^5.9.1",
     "@types/testing-library__react": "^9.1.2",
-    "@types/uuid": "^8.3.0",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
     "jest-when": "^3.5.0",
