+++ packages/react-hooks/src/useCancelableEffect.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { EffectCallback, useEffect } from "react";
+import { Holder } from "./Holder";
+
+export function useCancelableEffect(effect: (args: { canceled: Holder<boolean> }) => ReturnType<EffectCallback>) {
+  useEffect(() => {
+    const canceled = new Holder(false);
+
+    const effectCleanup = effect({ canceled });
+
+    return () => {
+      canceled.set(true);
+      if (effectCleanup) {
+        effectCleanup();
+      }
+    };
+  }, [effect]);
+}
