+++ packages/react-hooks/src/useController.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { useCallback, useState } from "react";
+
+export function useController<T>(): [T | undefined, (controller: T) => void] {
+  const [controller, setController] = useState<T | undefined>(undefined);
+
+  const ref = useCallback((controller: T) => {
+    setController(controller);
+  }, []);
+
+  return [controller, ref];
+}
