+++ repo/graph.dot
@@ -88,6 +88,7 @@ digraph G {
   "@kie-tools/unitables-dmn" [ color = "blue", fontcolor = "blue", style = "rounded" ];
   "@kie-tools/openjdk11-mvn-image" [ color = "black", fontcolor = "black", style = "dashed, rounded" ];
   "@kie-tools/openjdk11-mvn-image-env" [ color = "black", fontcolor = "black", style = "dashed, rounded" ];
+  "@kie-tools-core/react-hooks" [ color = "purple", fontcolor = "purple", style = "rounded" ];
   "@kie-tools/serverless-logic-sandbox" [ color = "black", fontcolor = "black", style = "dashed, rounded" ];
   "@kie-tools/text-editor" [ color = "blue", fontcolor = "blue", style = "rounded" ];
   "@kie-tools/serverless-logic-sandbox-base-image-env" [ color = "black", fontcolor = "black", style = "dashed, rounded" ];
@@ -276,6 +277,9 @@ digraph G {
   "@kie-tools/pmml-editor-marshaller" -> "@kie-tools/eslint" [ style = "dashed", color = "blue" ];
   "@kie-tools/pmml-editor-marshaller" -> "@kie-tools/root-env" [ style = "dashed", color = "blue" ];
   "@kie-tools/pmml-editor-marshaller" -> "@kie-tools/tsconfig" [ style = "dashed", color = "blue" ];
+  "@kie-tools-core/react-hooks" -> "@kie-tools/eslint" [ style = "dashed", color = "purple" ];
+  "@kie-tools-core/react-hooks" -> "@kie-tools/root-env" [ style = "dashed", color = "purple" ];
+  "@kie-tools-core/react-hooks" -> "@kie-tools/tsconfig" [ style = "dashed", color = "purple" ];
   "@kie-tools/serverless-logic-sandbox" -> "@kie-tools-core/workspaces-git-fs" [ style = "solid", color = "black" ];
   "@kie-tools/serverless-logic-sandbox" -> "@kie-tools/dashbuilder-editor" [ style = "solid", color = "black" ];
   "@kie-tools/serverless-logic-sandbox" -> "@kie-tools/i18n-common-dictionary" [ style = "solid", color = "black" ];
@@ -367,6 +371,7 @@ digraph G {
   "@kie-tools-core/webpack-base" -> "@kie-tools/root-env" [ style = "dashed", color = "black" ];
   "@kie-tools-core/workspace" -> "@kie-tools-core/operating-system" [ style = "solid", color = "purple" ];
   "@kie-tools-core/workspaces-git-fs" -> "@kie-tools-core/envelope-bus" [ style = "solid", color = "purple" ];
+  "@kie-tools-core/workspaces-git-fs" -> "@kie-tools-core/react-hooks" [ style = "solid", color = "purple" ];
   "@kie-tools-core/workspaces-git-fs" -> "@kie-tools-core/workspace" [ style = "solid", color = "purple" ];
   "@kie-tools-core/workspaces-git-fs" -> "@kie-tools/kie-sandbox-fs" [ style = "solid", color = "purple" ];
   "@kie-tools/yard-editor" -> "@kie-tools-core/editor" [ style = "solid", color = "blue" ];
