+++ repo/graph.json
@@ -91,6 +91,7 @@
       { "id": "@kie-tools/pmml-editor" },
       { "id": "@kie-tools/unitables-dmn" },
       { "id": "@kie-tools/kn-plugin-workflow" },
+      { "id": "@kie-tools-core/react-hooks" },
       { "id": "@kie-tools/unitables" },
       { "id": "@kie-tools/openjdk11-mvn-image" },
       { "id": "@kie-tools/openjdk11-mvn-image-env" },
@@ -545,6 +546,7 @@
       { "source": "@kie-tools/online-editor", "target": "@kie-tools/extended-services", "weight": 1 },
       { "source": "@kie-tools/online-editor", "target": "@kie-tools/stunner-editors", "weight": 1 },
       { "source": "@kie-tools-core/workspaces-git-fs", "target": "@kie-tools-core/envelope-bus", "weight": 1 },
+      { "source": "@kie-tools-core/workspaces-git-fs", "target": "@kie-tools-core/react-hooks", "weight": 1 },
       { "source": "@kie-tools-core/workspaces-git-fs", "target": "@kie-tools-core/workspace", "weight": 1 },
       { "source": "@kie-tools-core/workspaces-git-fs", "target": "@kie-tools/kie-sandbox-fs", "weight": 1 },
       { "source": "@kie-tools/pmml-editor", "target": "@kie-tools-core/editor", "weight": 1 },
@@ -554,6 +556,9 @@
       { "source": "@kie-tools/unitables-dmn", "target": "@kie-tools/form-dmn", "weight": 1 },
       { "source": "@kie-tools/unitables-dmn", "target": "@kie-tools/unitables", "weight": 1 },
       { "source": "@kie-tools/kn-plugin-workflow", "target": "@kie-tools/root-env", "weight": 1 },
+      { "source": "@kie-tools-core/react-hooks", "target": "@kie-tools/eslint", "weight": 1 },
+      { "source": "@kie-tools-core/react-hooks", "target": "@kie-tools/root-env", "weight": 1 },
+      { "source": "@kie-tools-core/react-hooks", "target": "@kie-tools/tsconfig", "weight": 1 },
       { "source": "@kie-tools/unitables", "target": "@kie-tools/boxed-expression-component", "weight": 1 },
       { "source": "@kie-tools/unitables", "target": "@kie-tools/form", "weight": 1 },
       { "source": "@kie-tools/openjdk11-mvn-image", "target": "@kie-tools/openjdk11-mvn-image-env", "weight": 1 },
@@ -782,6 +787,7 @@
     ["@kie-tools-core/patternfly-base", "packages/patternfly-base"],
     ["@kie-tools/pmml-editor", "packages/pmml-editor"],
     ["@kie-tools/pmml-editor-marshaller", "packages/pmml-editor-marshaller"],
+    ["@kie-tools-core/react-hooks", "packages/react-hooks"],
     ["@kie-tools/root-env", "packages/root-env"],
     ["@kie-tools/serverless-logic-sandbox", "packages/serverless-logic-sandbox"],
     ["@kie-tools/serverless-logic-sandbox-base-image", "packages/serverless-logic-sandbox-base-image"],
